backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email-validator]==2.5.2
textblob==0.17.1
python-dotenv==1.0.0
aiosqlite==0.19.0
websockets==12.0
bcrypt==4.0.1
google-generativeai
